requirements.txt
python-telegram-bot>=20.3
notion-client>=2.0.0
openai>=1.3.0
requests>=2.28.0